Models/requirements.txt
pandas
numpy
matplotlib 
Pillow
h5py
torchvision
tensorflow
tensorflow.keras
scikit-learn
monai
opencv-python